Move decision maker in lib
Not sure why it was part of externalities given it’s part of the
exposed API and is clearly related to Kite’s install
