+++ lib/decision-maker.js
@@ -2,9 +2,9 @@
 
 const os = require('os');
 
-const Client = require('../lib/client.js');
-const KiteError = require('../lib/kite-error.js');
-const utils = require('../lib/utils.js');
+const Client = require('./client.js');
+const KiteError = require('./kite-error.js');
+const utils = require('./utils.js');
 
 class DecisionMaker {
   constructor(editor, plugin) {
