+++ lib/index.js
@@ -1,7 +1,7 @@
 module.exports = {
   AccountManager: require('./account-manager.js'),
   AtomHelper: require('./atom-helper.js'),
-  DecisionMaker: require('../ext/decision-maker.js'),
+  DecisionMaker: require('./decision-maker.js'),
   Installation: require('./models/installation.js'),
   Installer: require('./installer.js'),
   InstallFlow: require('./elements/install-flow.js'),
